+++ src/Query/Contracts/ClassQueryInterface.php
@@ -9,32 +9,32 @@ interface ClassQueryInterface extends Query
     /**
      * @param  class-string  $classpath
      */
-    public function whereExtends(string $classpath): self;
+    public function whereExtends(string $classpath, bool $all = true): self;
 
     /**
      * @param  class-string  $classpath
      */
-    public function whereDoesntExtend(string $classpath): self;
+    public function whereDoesntExtend(string $classpath, bool $all = true): self;
 
     /**
      * @param  class-string|class-string[]  $interface
      */
-    public function whereImplements(string|array $interface): self;
+    public function whereImplements(string|array $interface, bool $all = true): self;
 
     /**
      * @param  class-string|class-string[]  $interface
      */
-    public function whereDoesntImplement(string|array $interface): self;
+    public function whereDoesntImplement(string|array $interface, bool $all = true): self;
 
     /**
      * @param  class-string|class-string[]  $trait
      */
-    public function whereUses(string|array $trait): self;
+    public function whereUses(string|array $trait, bool $all = true): self;
 
     /**
      * @param  class-string|class-string[]  $trait
      */
-    public function whereDoesntUse(string|array $trait): self;
+    public function whereDoesntUse(string|array $trait, bool $all = true): self;
 
     public function whereNameContains(string|array $text): self;
     public function whereNameDoesntContain(string|array $text, bool $all = true): self;