+++ src/Query/Contracts/RouteQueryInterface.php
@@ -7,6 +7,8 @@ use Mateffy\Introspect\Query\Query;
 interface RouteQueryInterface extends Query
 {
     // Route names
+    public function whereNameEquals(string $name): static;
+    public function whereNameDoesntEqual(string $name): static;
     public function whereNameStartsWith(string $text): static;
 
     public function whereNameDoesntStartWith(string $text): static;