+++ src/Query/Where/Classes/WhereClassNameContains.php
@@ -4,6 +4,7 @@ namespace Mateffy\Introspect\Query\Where\Classes;
 
 use Mateffy\Introspect\Query\Where\Generic\WhereTextContains;
 use Mateffy\Introspect\Query\Where\RouteWhere;
+use Mateffy\Introspect\Reflection\ModelReflector;
 use ReflectionClass;
 
 class WhereClassNameContains implements RouteWhere
@@ -11,7 +12,7 @@ class WhereClassNameContains implements RouteWhere
     use WhereTextContains;
 
     /**
-     * @param  ReflectionClass  $value
+     * @param  ReflectionClass|ModelReflector  $value
      */
     protected function getName($value): ?string
     {