+++ src/Query/Where/Classes/WhereClassNameEquals.php
@@ -4,6 +4,7 @@ namespace Mateffy\Introspect\Query\Where\Classes;
 
 use Mateffy\Introspect\Query\Where\Generic\WhereTextEquals;
 use Mateffy\Introspect\Query\Where\RouteWhere;
+use Mateffy\Introspect\Reflection\ModelReflector;
 use ReflectionClass;
 
 class WhereClassNameEquals implements RouteWhere
@@ -11,7 +12,7 @@ class WhereClassNameEquals implements RouteWhere
     use WhereTextEquals;
 
     /**
-     * @param  ReflectionClass  $value
+     * @param  ReflectionClass|ModelReflector  $value
      */
     protected function getName($value): ?string
     {